refactor(reth): retire JSONL path; cgo direct-write is the only path
Delete the JSONL+subprocess implementation now that the cgo direct-write
path (Slices A-E) boots stock reth via TestRethNodeBoot:

  - populate.go (243 LOC) — the JSONL Populate function
  - populate_test.go (407 LOC)
  - statedump.go (279 LOC) — JSONL streaming + StackTrie root
  - reth_binary.go (82 LOC) — reth subprocess management +
    verifyRethVersion allowlist
  - entities.go (237 LOC) — RNG-driven entity gen (replaced by
    internal/entitygen)

Trimmed:
  - header.go: dropped buildBlock1Header (the v2-underflow workaround
    was JSONL-only) and writeHeaderFile (wrote RLP for reth's --header
    flag); kept buildBlock0Header / buildHeaderFromGenesis /
    defaultGenesisForReth / ptrU64 (used by RunCgo).
  - doc.go: rewritten to describe the cgo direct-write architecture
    instead of the JSONL approach.

Created:
  - options.go: hosts the Options struct (now empty, future-reserved),
    the GenesisFilePath / ChainIDOverride package globals, and the
    genesisPathFromCfg / chainIDFromCfg trampolines that previously
    lived in populate.go.

Updated:
  - main.go: --client=reth now dispatches to reth.RunCgo. Without
    -tags cgo_reth, RunCgo returns errNotImplemented pointing at
    Dockerfile.reth — local builds without libmdbx + librocksdb cannot
    exercise the path, by design.

The Slice F differential oracle (JSONL-vs-cgo equivalence) goal becomes
moot — there is no JSONL path to compare against. The remaining Slice F
work (deprecation timeline, retire-via-stages doc, etc.) is also
collapsed: the retirement happens in this commit.
